Refactor Settings to be more modular
